Update examples
Add copyright to example 2 and example 3.
